Review debug_traceBlochByNumber, debug_traceBlochByHsh, debug_traceTransaction and debug_traceCall benches for the sake of clarity
